# openapi.yaml
openapi: 3.1.0
info:
  title: Codestra Face Liveness API
  description: 'Standalone passive presentation-attack-detection (PAD) service for Codestra FACE-ID. It
    answers one question: is the single face in this image a bona fide live capture or a presentation
    attack (print, screen replay, mask)? It does NOT perform face matching or identity verification. The
    only supported caller is Middleware V3, which owns authentication normalisation, policy, orchestration
    and cross-system audit. Images are processed in memory and never persisted.'
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Local development
paths:
...
  /v1/capabilities:
    get:
      tags:
      - liveness
      summary: Capabilities
      description: What this service can and cannot do, its limits, and its decision policy.
      operationId: capabilities
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilitiesResponse'
        '401':
          description: 'Error codes: UNAUTHORIZED'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ArtifactOut:
      properties:
        name:
          type: string
          title: Name
        file:
          type: string
          title: File
        sha256:
          type: string
          title: Sha256
      type: object
      required:
      - name
      - file
      - sha256
      title: ArtifactOut
    CapabilitiesResponse:
      properties:
        service:
          type: string
          title: Service
        version:
          type: string
          title: Version
        api_version:
          type: string
          const: v1
          title: Api Version
          default: v1
        ready:
          type: boolean
          title: Ready
        methods:
          items:
            $ref: '#/components/schemas/MethodInfo'
          type: array
          title: Methods
        unsupported:
          items:
            type: string
          type: array
          title: Unsupported
        input:
          $ref: '#/components/schemas/InputLimits'
        decision:
          $ref: '#/components/schemas/DecisionPolicy'
        model:
          $ref: '#/components/schemas/ModelInfo'
        face_matching:
          type: boolean
          const: false
          title: Face Matching
          default: false
        image_persistence:
          type: string
          const: none
          title: Image Persistence
          default: none
      type: object
      required:
      - service
      - version
      - ready
      - methods
      - unsupported
      - input
      - decision
      - model
      title: CapabilitiesResponse
...
    DecisionPolicy:
      properties:
        score:
          type: string
          title: Score
        threshold:
          type: number
          title: Threshold
        threshold_calibrated:
          type: boolean
          title: Threshold Calibrated
        calibration_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Calibration Id
      type: object
      required:
      - score
      - threshold
      - threshold_calibrated
      - calibration_id
      title: DecisionPolicy
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
          description: Stable machine-readable error code.
        message:
          type: string
          title: Message
        retryable:
          type: boolean
          title: Retryable
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      type: object
      required:
      - code
      - message
      - retryable
      - request_id
      title: ErrorBody
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - error
      title: ErrorResponse
...
    InputLimits:
      properties:
        formats:
          items:
            type: string
          type: array
          title: Formats
        max_image_bytes:
          type: integer
          title: Max Image Bytes
        max_image_pixels:
          type: integer
          title: Max Image Pixels
        min_image_side_px:
          type: integer
          title: Min Image Side Px
        min_face_size_px:
          type: integer
          title: Min Face Size Px
        faces_required:
          type: integer
          const: 1
          title: Faces Required
          default: 1
      type: object
      required:
      - formats
      - max_image_bytes
      - max_image_pixels
      - min_image_side_px
      - min_face_size_px
      title: InputLimits
...
    MethodInfo:
      properties:
        id:
          type: string
          title: Id
        description:
          type: string
          title: Description
      type: object
      required:
      - id
      - description
      title: MethodInfo
    ModelInfo:
      properties:
        id:
          type: string
          title: Id
        detector:
          type: string
          title: Detector
        license:
          type: string
          title: License
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactOut'
          type: array
          title: Artifacts
      type: object
      required:
      - id
      - detector
      - license
      - artifacts
      title: ModelInfo
...